+++ frontend/src/components/layout/AppFooter.vue
