+++ frontend/src/components/layout/AppNavbar.vue
