+++ frontend/src/components/users/PictureInput.vue
