+++ frontend/src/views/users/SignOut.vue
